tsweb: remove JSONHandlerFunc
It's unused and we've decided it's not what we want.
