+++ tsweb/tsweb.go
@@ -24,6 +24,7 @@ import (
 	"strings"
 	"time"
 
+	"go4.org/mem"
 	"inet.af/netaddr"
 	"tailscale.com/envknob"
 	"tailscale.com/metrics"
@@ -82,6 +83,30 @@ func AllowDebugAccess(r *http.Request) bool {
 			if err == nil && string(bytes.TrimSpace(slurp)) == urlKey {
 				return true
 			}
+		}
+	}
+	return false
+}
+
+
+// AcceptsEncoding reports whether r accepts the named encoding
+// ("gzip", "br", etc).
+func AcceptsEncoding(r *http.Request, enc string) bool {
+	h := r.Header.Get("Accept-Encoding")
+	if h == "" {
+		return false
+	}
+	if !strings.Contains(h, enc) && !mem.ContainsFold(mem.S(h), mem.S(enc)) {
+		return false
+	}
+	remain := h
+	for len(remain) > 0 {
+		var part string
+		part, remain, _ = strings.Cut(remain, ",")
+		part = strings.TrimSpace(part)
+		part, _, _ = strings.Cut(part, ";")
+		if part == enc {
+			return true
 		}
 	}
 	return false
